refactor directory to separate tests and create a shoots module
