+++ shoots/__init__.py
@@ -1,0 +1,8 @@
+from .shoots_server import ShootsServer
+from .shoots_client import ShootsClient, PutMode, BucketDeleteMode, JWTClientAuthHandler
+
+__all__ = ['ShootsServer', 
+           'ShootsClient', 
+           'PutMode', 
+           'BucketDeleteMode',
+           'JWTClientAuthHandler']
