+++ tests/tls_test.py
@@ -1,6 +1,5 @@
-from base_test import BaseTest
-from shoots_server import ShootsServer
-from shoots_client import ShootsClient
+from .base_test import BaseTest
+from shoots import ShootsServer, ShootsClient
 from pyarrow.flight import Location, FlightClient, FlightUnavailableError
 
 class TLSTest(BaseTest):
